requirements.txt
Flask==1.1.1
psycopg2==2.8.4
SQLAlchemy==1.3.15
marshmallow==3.5.1
gunicorn==20.0.4
jwt==0.6.1
python-dotenv==0.12.0
python-telegram-bot==12.4.2
